+++ tests/conftest.py
@@ -1,6 +1,9 @@
 # Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2022 CVAT.ai Corporation
 #
 # SPDX-License-Identifier: MIT
+
+from .fixtures import *
 
 
 def pytest_configure(config):